financial_research_assistant/pyproject.toml
[project]
name = "financial-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Fahad Khan", email = "fghory@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.11",
    "pydantic>=2.11.3",
    "rich>=14.0.0",
    "tavily-python>=0.5.4",
]

[project.scripts]
financial-research-assistant = "financial_research_assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
